backend/requirements-ai.txt
# Basic dependencies
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
redis==4.6.0
gunicorn==21.2.0

# AI/ML dependencies
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
python-Levenshtein==0.21.1
fuzzywuzzy==0.18.0

# Optional: Advanced ML (can be added later)
# sentence-transformers==2.2.2
# transformers==4.21.3
